context/feat: tokenizersetter / mv parts
