+++ context/errors.go
@@ -19,7 +19,9 @@ var (
 	ErrUserPromptEmpty   = errors.New("user prompt is empty")
 	ErrTokenizerNotFound = errors.New("tokenizer not found")
 
-	ErrMessageNotFound      = errors.New("message not found")
-	ErrContentUnmarshal     = errors.New("failed to unmarshal content type")
-	ErrUnknownContentType   = errors.New("unknown content type")
+	ErrMessageNotFound    = errors.New("message not found")
+	ErrContentUnmarshal   = errors.New("failed to unmarshal content type")
+	ErrUnknownContentType = errors.New("unknown content type")
+
+	ErrInvalidTokenLimit = errors.New("invalid token limit")
 )
